remove CustomCommandHandler and add before method in LilirucaCommand
