+++ src/structures/LilirucaCommand.js
@@ -1,6 +1,8 @@
 const { Command } = require('discord-akairo')
 const logger = require('@utils/logger')
 const locales = require('@utils/locales')
+const { DEFAULT_PREFIX, DEFAULT_LANGUAGE } = require('../Constants')
+
 /**
  * @property {import('@LilirucaClient')} client
  */
@@ -34,6 +36,23 @@ class LilirucaCommand extends Command {
         this.logger.warn('not found usage.')
       }
     }
+  }
+
+  async before (message) {
+    const guildData = await this.client.db.guilds.get(message.guild.id)
+    const language = guildData.language || DEFAULT_LANGUAGE
+    const t = this.client.locales.getT(language)
+    const ct = (tPath, tOptions) => t(`commands:${this.id}.${tPath}`, tOptions)
+
+    message.guildData = guildData
+    message.t = t
+    message.ct = ct
+    message.language = language
+    message.locales = this.client.locales
+    message.db = this.client.db
+    message.prefix = guildData.prefix || DEFAULT_PREFIX
+
+    return message
   }
 }
