+++ src/structures/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   LilirucaCollection: require('./LilirucaCollection'),
   LilirucaCommand: require('./LilirucaCommand'),
-  CustomCommandHandler: require('./CustomCommandHandler'),
   LilirucaEmbed: require('./LilirucaEmbed')
 }
